+++ src/wai/common/file/__init__.py
@@ -1,2 +1,4 @@
 from ._NamedColumnSelection import NamedColumnSelection, SELECTION_TYPE
 from ._functions import load_dir
+from ._PathContextManager import PathContextManager
+from ._util import ensure_path